+++ versions.tf
@@ -1,8 +1,8 @@
 terraform {
-  required_version = ">= 0.12.2"
+  required_version = ">= 0.12.6"
 
   required_providers {
-    aws      = ">= 2.31.0"
+    aws      = ">= 2.38.0"
     local    = ">= 1.2"
     null     = ">= 2.1"
     template = ">= 2.1"